Split frontend/backend models into a models.proto
