+++ console/src/gen/openauth/backend/v1/models_pb.ts
@@ -1,0 +1,882 @@
+// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
+// @generated from file openauth/backend/v1/models.proto (package openauth.backend.v1, syntax proto3)
+/* eslint-disable */
+
+import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
+import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
+import type { Timestamp } from "@bufbuild/protobuf/wkt";
+import { file_google_protobuf_struct, file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
+import type { JsonObject, Message } from "@bufbuild/protobuf";
+
+/**
+ * Describes the file openauth/backend/v1/models.proto.
+ */
+export const file_openauth_backend_v1_models: GenFile = /*@__PURE__*/
+  fileDesc("CiBvcGVuYXV0aC9iYWNrZW5kL3YxL21vZGVscy5wcm90bxITb3BlbmF1dGguYmFja2VuZC52MSLdBgoHUHJvamVjdBIKCgJpZBgBIAEoCRIUCgxkaXNwbGF5X25hbWUYAiABKAkSLwoLY3JlYXRlX3RpbWUYAyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEi8KC3VwZGF0ZV90aW1lGAQgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIfChJsb2dfaW5fd2l0aF9nb29nbGUYBiABKAhIAIgBARIiChVsb2dfaW5fd2l0aF9taWNyb3NvZnQYByABKAhIAYgBARIeChFsb2dfaW5fd2l0aF9lbWFpbBgPIAEoCEgCiAEBEiEKFGxvZ19pbl93aXRoX3Bhc3N3b3JkGAUgASgISAOIAQESHQoQbG9nX2luX3dpdGhfc2FtbBgQIAEoCEgEiAEBEioKHWxvZ19pbl93aXRoX2F1dGhlbnRpY2F0b3JfYXBwGA0gASgISAWIAQESIAoTbG9nX2luX3dpdGhfcGFzc2tleRgOIAEoCEgGiAEBEh4KFmdvb2dsZV9vYXV0aF9jbGllbnRfaWQYCCABKAkSIgoaZ29vZ2xlX29hdXRoX2NsaWVudF9zZWNyZXQYCSABKAkSIQoZbWljcm9zb2Z0X29hdXRoX2NsaWVudF9pZBgKIAEoCRIlCh1taWNyb3NvZnRfb2F1dGhfY2xpZW50X3NlY3JldBgLIAEoCRIYCgthdXRoX2RvbWFpbhgMIAEoCUgHiAEBEhcKD3RydXN0ZWRfZG9tYWlucxgRIAMoCRIUCgxyZWRpcmVjdF91cmkYEiABKAkSIAoYYWZ0ZXJfbG9naW5fcmVkaXJlY3RfdXJpGBMgASgJEiEKGWFmdGVyX3NpZ251cF9yZWRpcmVjdF91cmkYFCABKAlCFQoTX2xvZ19pbl93aXRoX2dvb2dsZUIYChZfbG9nX2luX3dpdGhfbWljcm9zb2Z0QhQKEl9sb2dfaW5fd2l0aF9lbWFpbEIXChVfbG9nX2luX3dpdGhfcGFzc3dvcmRCEwoRX2xvZ19pbl93aXRoX3NhbWxCIAoeX2xvZ19pbl93aXRoX2F1dGhlbnRpY2F0b3JfYXBwQhYKFF9sb2dfaW5fd2l0aF9wYXNza2V5Qg4KDF9hdXRoX2RvbWFpbiLrAgoTVmF1bHREb21haW5TZXR0aW5ncxIWCg5wZW5kaW5nX2RvbWFpbhgBIAEoCRJQChV0ZXNzZXJhbF9jbmFtZV9yZWNvcmQYAiABKAsyMS5vcGVuYXV0aC5iYWNrZW5kLnYxLlZhdWx0RG9tYWluU2V0dGluZ3NETlNSZWNvcmQSKAogdGVzc2VyYWxfY25hbWVfcmVjb3JkX2NvbmZpZ3VyZWQYAyABKAgSRwoMZGtpbV9yZWNvcmRzGAQgAygLMjEub3BlbmF1dGguYmFja2VuZC52MS5WYXVsdERvbWFpblNldHRpbmdzRE5TUmVjb3JkEhcKD2RraW1fY29uZmlndXJlZBgFIAEoCBJGCgtzcGZfcmVjb3JkcxgGIAMoCzIxLm9wZW5hdXRoLmJhY2tlbmQudjEuVmF1bHREb21haW5TZXR0aW5nc0ROU1JlY29yZBIWCg5zcGZfY29uZmlndXJlZBgHIAEoCCJJChxWYXVsdERvbWFpblNldHRpbmdzRE5TUmVjb3JkEgwKBHR5cGUYASABKAkSDAoEbmFtZRgCIAEoCRINCgV2YWx1ZRgDIAEoCSLJAgoRUHJvamVjdFVJU2V0dGluZ3MSCgoCaWQYASABKAkSEgoKcHJvamVjdF9pZBgCIAEoCRIvCgtjcmVhdGVfdGltZRgDIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYBCABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEhAKCGxvZ29fdXJsGAUgASgJEhMKC2Zhdmljb25fdXJsGAYgASgJEhUKDXByaW1hcnlfY29sb3IYByABKAkSIAoYZGV0ZWN0X2RhcmtfbW9kZV9lbmFibGVkGAggASgIEhoKEmRhcmtfbW9kZV9sb2dvX3VybBgJIAEoCRIfChdkYXJrX21vZGVfcHJpbWFyeV9jb2xvchgKIAEoCRIVCg1sb2dfaW5fbGF5b3V0GAsgASgJIowFCgxPcmdhbml6YXRpb24SCgoCaWQYASABKAkSFAoMZGlzcGxheV9uYW1lGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgt1cGRhdGVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASHwoSbG9nX2luX3dpdGhfZ29vZ2xlGAcgASgISACIAQESIgoVbG9nX2luX3dpdGhfbWljcm9zb2Z0GAggASgISAGIAQESHgoRbG9nX2luX3dpdGhfZW1haWwYDiABKAhIAogBARIhChRsb2dfaW5fd2l0aF9wYXNzd29yZBgGIAEoCEgDiAEBEh0KEGxvZ19pbl93aXRoX3NhbWwYCSABKAhIBIgBARIqCh1sb2dfaW5fd2l0aF9hdXRoZW50aWNhdG9yX2FwcBgLIAEoCEgFiAEBEiAKE2xvZ19pbl93aXRoX3Bhc3NrZXkYDCABKAhIBogBARIYCgtyZXF1aXJlX21mYRgNIAEoCEgHiAEBEhkKDHNjaW1fZW5hYmxlZBgKIAEoCEgIiAEBQhUKE19sb2dfaW5fd2l0aF9nb29nbGVCGAoWX2xvZ19pbl93aXRoX21pY3Jvc29mdEIUChJfbG9nX2luX3dpdGhfZW1haWxCFwoVX2xvZ19pbl93aXRoX3Bhc3N3b3JkQhMKEV9sb2dfaW5fd2l0aF9zYW1sQiAKHl9sb2dfaW5fd2l0aF9hdXRoZW50aWNhdG9yX2FwcEIWChRfbG9nX2luX3dpdGhfcGFzc2tleUIOCgxfcmVxdWlyZV9tZmFCDwoNX3NjaW1fZW5hYmxlZCJZCh9Pcmdhbml6YXRpb25Hb29nbGVIb3N0ZWREb21haW5zEhcKD29yZ2FuaXphdGlvbl9pZBgBIAEoCRIdChVnb29nbGVfaG9zdGVkX2RvbWFpbnMYAiADKAkiVwoeT3JnYW5pemF0aW9uTWljcm9zb2Z0VGVuYW50SURzEhcKD29yZ2FuaXphdGlvbl9pZBgBIAEoCRIcChRtaWNyb3NvZnRfdGVuYW50X2lkcxgCIAMoCSJQChFTZXNzaW9uU2lnbmluZ0tleRIKCgJpZBgBIAEoCRIvCg5wdWJsaWNfa2V5X2p3axgCIAEoCzIXLmdvb2dsZS5wcm90b2J1Zi5TdHJ1Y3QiugEKDVByb2plY3RBUElLZXkSCgoCaWQYASABKAkSFAoMZGlzcGxheV9uYW1lGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgt1cGRhdGVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASFAoMc2VjcmV0X3Rva2VuGAUgASgJEg8KB3Jldm9rZWQYBiABKAgilAEKDlB1Ymxpc2hhYmxlS2V5EgoKAmlkGAEgASgJEhQKDGRpc3BsYXlfbmFtZRgCIAEoCRIvCgtjcmVhdGVfdGltZRgDIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYBCABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wIq8BChNJbnRlcm1lZGlhdGVTZXNzaW9uEgoKAmlkGAEgASgJEg0KBWVtYWlsGAIgASgJEhwKFGdvb2dsZV9ob3N0ZWRfZG9tYWluGAMgASgJEhYKDmdvb2dsZV91c2VyX2lkGAQgASgJEhsKE21pY3Jvc29mdF90ZW5hbnRfaWQYBSABKAkSGQoRbWljcm9zb2Z0X3VzZXJfaWQYBiABKAkSDwoHcmV2b2tlZBgHIAEoCCK/AgoEVXNlchIKCgJpZBgBIAEoCRIXCg9vcmdhbml6YXRpb25faWQYAiABKAkSDQoFZW1haWwYAyABKAkSLwoLY3JlYXRlX3RpbWUYBCABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEi8KC3VwZGF0ZV90aW1lGAUgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBISCgVvd25lchgGIAEoCEgAiAEBEhsKDmdvb2dsZV91c2VyX2lkGAcgASgJSAGIAQESHgoRbWljcm9zb2Z0X3VzZXJfaWQYCCABKAlIAogBARIdChVoYXNfYXV0aGVudGljYXRvcl9hcHAYCSABKAhCCAoGX293bmVyQhEKD19nb29nbGVfdXNlcl9pZEIUChJfbWljcm9zb2Z0X3VzZXJfaWQizwEKB1Nlc3Npb24SCgoCaWQYASABKAkSDwoHdXNlcl9pZBgCIAEoCRIPCgdyZXZva2VkGAMgASgIEi8KC2NyZWF0ZV90aW1lGAQgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBI0ChBsYXN0X2FjdGl2ZV90aW1lGAUgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgtleHBpcmVfdGltZRgGIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXAi+wEKB1Bhc3NrZXkSCgoCaWQYASABKAkSDwoHdXNlcl9pZBgCIAEoCRIvCgtjcmVhdGVfdGltZRgDIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYBCABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEhUKCGRpc2FibGVkGAggASgISACIAQESFQoNY3JlZGVudGlhbF9pZBgFIAEoDBIXCg9wdWJsaWNfa2V5X3BraXgYBiABKAkSDgoGYWFndWlkGAcgASgJEg0KBXJwX2lkGAkgASgJQgsKCV9kaXNhYmxlZCKyAgoOU0FNTENvbm5lY3Rpb24SCgoCaWQYASABKAkSFwoPb3JnYW5pemF0aW9uX2lkGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgt1cGRhdGVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASFAoHcHJpbWFyeRgFIAEoCEgAiAEBEhIKCnNwX2Fjc191cmwYBiABKAkSFAoMc3BfZW50aXR5X2lkGAcgASgJEhgKEGlkcF9yZWRpcmVjdF91cmwYCCABKAkSHAoUaWRwX3g1MDlfY2VydGlmaWNhdGUYCSABKAkSFQoNaWRwX2VudGl0eV9pZBgKIAEoCUIKCghfcHJpbWFyeSLQAQoKU0NJTUFQSUtleRIKCgJpZBgBIAEoCRIXCg9vcmdhbml6YXRpb25faWQYAiABKAkSLwoLY3JlYXRlX3RpbWUYAyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEi8KC3VwZGF0ZV90aW1lGAQgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIUCgxkaXNwbGF5X25hbWUYBSABKAkSFAoMc2VjcmV0X3Rva2VuGAYgASgJEg8KB3Jldm9rZWQYByABKAgizgEKFlVzZXJJbXBlcnNvbmF0aW9uVG9rZW4SCgoCaWQYASABKAkSFwoPaW1wZXJzb25hdG9yX2lkGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgtleHBpcmVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASFAoMc2VjcmV0X3Rva2VuGAUgASgJEhcKD2ltcGVyc29uYXRlZF9pZBgGIAEoCSKxAQoKVXNlckludml0ZRIKCgJpZBgBIAEoCRIXCg9vcmdhbml6YXRpb25faWQYAiABKAkSLwoLY3JlYXRlX3RpbWUYAyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEi8KC3VwZGF0ZV90aW1lGAQgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBINCgVlbWFpbBgFIAEoCRINCgVvd25lchgGIAEoCELlAQoXY29tLm9wZW5hdXRoLmJhY2tlbmQudjFCC01vZGVsc1Byb3RvUAFaT2dpdGh1Yi5jb20vb3BlbmF1dGgvb3BlbmF1dGgvaW50ZXJuYWwvYmFja2VuZC9nZW4vb3BlbmF1dGgvYmFja2VuZC92MTtiYWNrZW5kdjGiAgNPQliqAhNPcGVuYXV0aC5CYWNrZW5kLlYxygITT3BlbmF1dGhcQmFja2VuZFxWMeICH09wZW5hdXRoXEJhY2tlbmRcVjFcR1BCTWV0YWRhdGHqAhVPcGVuYXV0aDo6QmFja2VuZDo6VjFiBnByb3RvMw", [file_google_protobuf_timestamp, file_google_protobuf_struct]);
+
+/**
+ * @generated from message openauth.backend.v1.Project
+ */
+export type Project = Message<"openauth.backend.v1.Project"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool log_in_with_google = 6;
+   */
+  logInWithGoogle?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_microsoft = 7;
+   */
+  logInWithMicrosoft?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_email = 15;
+   */
+  logInWithEmail?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_password = 5;
+   */
+  logInWithPassword?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_saml = 16;
+   */
+  logInWithSaml?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_authenticator_app = 13;
+   */
+  logInWithAuthenticatorApp?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_passkey = 14;
+   */
+  logInWithPasskey?: boolean;
+
+  /**
+   * @generated from field: string google_oauth_client_id = 8;
+   */
+  googleOauthClientId: string;
+
+  /**
+   * @generated from field: string google_oauth_client_secret = 9;
+   */
+  googleOauthClientSecret: string;
+
+  /**
+   * @generated from field: string microsoft_oauth_client_id = 10;
+   */
+  microsoftOauthClientId: string;
+
+  /**
+   * @generated from field: string microsoft_oauth_client_secret = 11;
+   */
+  microsoftOauthClientSecret: string;
+
+  /**
+   * @generated from field: optional string auth_domain = 12;
+   */
+  authDomain?: string;
+
+  /**
+   * @generated from field: repeated string trusted_domains = 17;
+   */
+  trustedDomains: string[];
+
+  /**
+   * @generated from field: string redirect_uri = 18;
+   */
+  redirectUri: string;
+
+  /**
+   * @generated from field: string after_login_redirect_uri = 19;
+   */
+  afterLoginRedirectUri: string;
+
+  /**
+   * @generated from field: string after_signup_redirect_uri = 20;
+   */
+  afterSignupRedirectUri: string;
+};
+
+/**
+ * Describes the message openauth.backend.v1.Project.
+ * Use `create(ProjectSchema)` to create a new message.
+ */
+export const ProjectSchema: GenMessage<Project> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 0);
+
+/**
+ * @generated from message openauth.backend.v1.VaultDomainSettings
+ */
+export type VaultDomainSettings = Message<"openauth.backend.v1.VaultDomainSettings"> & {
+  /**
+   * @generated from field: string pending_domain = 1;
+   */
+  pendingDomain: string;
+
+  /**
+   * @generated from field: openauth.backend.v1.VaultDomainSettingsDNSRecord tesseral_cname_record = 2;
+   */
+  tesseralCnameRecord?: VaultDomainSettingsDNSRecord;
+
+  /**
+   * @generated from field: bool tesseral_cname_record_configured = 3;
+   */
+  tesseralCnameRecordConfigured: boolean;
+
+  /**
+   * @generated from field: repeated openauth.backend.v1.VaultDomainSettingsDNSRecord dkim_records = 4;
+   */
+  dkimRecords: VaultDomainSettingsDNSRecord[];
+
+  /**
+   * @generated from field: bool dkim_configured = 5;
+   */
+  dkimConfigured: boolean;
+
+  /**
+   * @generated from field: repeated openauth.backend.v1.VaultDomainSettingsDNSRecord spf_records = 6;
+   */
+  spfRecords: VaultDomainSettingsDNSRecord[];
+
+  /**
+   * @generated from field: bool spf_configured = 7;
+   */
+  spfConfigured: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.VaultDomainSettings.
+ * Use `create(VaultDomainSettingsSchema)` to create a new message.
+ */
+export const VaultDomainSettingsSchema: GenMessage<VaultDomainSettings> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 1);
+
+/**
+ * @generated from message openauth.backend.v1.VaultDomainSettingsDNSRecord
+ */
+export type VaultDomainSettingsDNSRecord = Message<"openauth.backend.v1.VaultDomainSettingsDNSRecord"> & {
+  /**
+   * @generated from field: string type = 1;
+   */
+  type: string;
+
+  /**
+   * @generated from field: string name = 2;
+   */
+  name: string;
+
+  /**
+   * @generated from field: string value = 3;
+   */
+  value: string;
+};
+
+/**
+ * Describes the message openauth.backend.v1.VaultDomainSettingsDNSRecord.
+ * Use `create(VaultDomainSettingsDNSRecordSchema)` to create a new message.
+ */
+export const VaultDomainSettingsDNSRecordSchema: GenMessage<VaultDomainSettingsDNSRecord> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 2);
+
+/**
+ * @generated from message openauth.backend.v1.ProjectUISettings
+ */
+export type ProjectUISettings = Message<"openauth.backend.v1.ProjectUISettings"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string project_id = 2;
+   */
+  projectId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: string logo_url = 5;
+   */
+  logoUrl: string;
+
+  /**
+   * @generated from field: string favicon_url = 6;
+   */
+  faviconUrl: string;
+
+  /**
+   * @generated from field: string primary_color = 7;
+   */
+  primaryColor: string;
+
+  /**
+   * @generated from field: bool detect_dark_mode_enabled = 8;
+   */
+  detectDarkModeEnabled: boolean;
+
+  /**
+   * @generated from field: string dark_mode_logo_url = 9;
+   */
+  darkModeLogoUrl: string;
+
+  /**
+   * @generated from field: string dark_mode_primary_color = 10;
+   */
+  darkModePrimaryColor: string;
+
+  /**
+   * @generated from field: string log_in_layout = 11;
+   */
+  logInLayout: string;
+};
+
+/**
+ * Describes the message openauth.backend.v1.ProjectUISettings.
+ * Use `create(ProjectUISettingsSchema)` to create a new message.
+ */
+export const ProjectUISettingsSchema: GenMessage<ProjectUISettings> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 3);
+
+/**
+ * @generated from message openauth.backend.v1.Organization
+ */
+export type Organization = Message<"openauth.backend.v1.Organization"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool log_in_with_google = 7;
+   */
+  logInWithGoogle?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_microsoft = 8;
+   */
+  logInWithMicrosoft?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_email = 14;
+   */
+  logInWithEmail?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_password = 6;
+   */
+  logInWithPassword?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_saml = 9;
+   */
+  logInWithSaml?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_authenticator_app = 11;
+   */
+  logInWithAuthenticatorApp?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_passkey = 12;
+   */
+  logInWithPasskey?: boolean;
+
+  /**
+   * @generated from field: optional bool require_mfa = 13;
+   */
+  requireMfa?: boolean;
+
+  /**
+   * @generated from field: optional bool scim_enabled = 10;
+   */
+  scimEnabled?: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.Organization.
+ * Use `create(OrganizationSchema)` to create a new message.
+ */
+export const OrganizationSchema: GenMessage<Organization> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 4);
+
+/**
+ * @generated from message openauth.backend.v1.OrganizationGoogleHostedDomains
+ */
+export type OrganizationGoogleHostedDomains = Message<"openauth.backend.v1.OrganizationGoogleHostedDomains"> & {
+  /**
+   * @generated from field: string organization_id = 1;
+   */
+  organizationId: string;
+
+  /**
+   * @generated from field: repeated string google_hosted_domains = 2;
+   */
+  googleHostedDomains: string[];
+};
+
+/**
+ * Describes the message openauth.backend.v1.OrganizationGoogleHostedDomains.
+ * Use `create(OrganizationGoogleHostedDomainsSchema)` to create a new message.
+ */
+export const OrganizationGoogleHostedDomainsSchema: GenMessage<OrganizationGoogleHostedDomains> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 5);
+
+/**
+ * @generated from message openauth.backend.v1.OrganizationMicrosoftTenantIDs
+ */
+export type OrganizationMicrosoftTenantIDs = Message<"openauth.backend.v1.OrganizationMicrosoftTenantIDs"> & {
+  /**
+   * @generated from field: string organization_id = 1;
+   */
+  organizationId: string;
+
+  /**
+   * @generated from field: repeated string microsoft_tenant_ids = 2;
+   */
+  microsoftTenantIds: string[];
+};
+
+/**
+ * Describes the message openauth.backend.v1.OrganizationMicrosoftTenantIDs.
+ * Use `create(OrganizationMicrosoftTenantIDsSchema)` to create a new message.
+ */
+export const OrganizationMicrosoftTenantIDsSchema: GenMessage<OrganizationMicrosoftTenantIDs> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 6);
+
+/**
+ * @generated from message openauth.backend.v1.SessionSigningKey
+ */
+export type SessionSigningKey = Message<"openauth.backend.v1.SessionSigningKey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: google.protobuf.Struct public_key_jwk = 2;
+   */
+  publicKeyJwk?: JsonObject;
+};
+
+/**
+ * Describes the message openauth.backend.v1.SessionSigningKey.
+ * Use `create(SessionSigningKeySchema)` to create a new message.
+ */
+export const SessionSigningKeySchema: GenMessage<SessionSigningKey> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 7);
+
+/**
+ * @generated from message openauth.backend.v1.ProjectAPIKey
+ */
+export type ProjectAPIKey = Message<"openauth.backend.v1.ProjectAPIKey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: string secret_token = 5;
+   */
+  secretToken: string;
+
+  /**
+   * @generated from field: bool revoked = 6;
+   */
+  revoked: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.ProjectAPIKey.
+ * Use `create(ProjectAPIKeySchema)` to create a new message.
+ */
+export const ProjectAPIKeySchema: GenMessage<ProjectAPIKey> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 8);
+
+/**
+ * @generated from message openauth.backend.v1.PublishableKey
+ */
+export type PublishableKey = Message<"openauth.backend.v1.PublishableKey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+};
+
+/**
+ * Describes the message openauth.backend.v1.PublishableKey.
+ * Use `create(PublishableKeySchema)` to create a new message.
+ */
+export const PublishableKeySchema: GenMessage<PublishableKey> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 9);
+
+/**
+ * @generated from message openauth.backend.v1.IntermediateSession
+ */
+export type IntermediateSession = Message<"openauth.backend.v1.IntermediateSession"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string email = 2;
+   */
+  email: string;
+
+  /**
+   * @generated from field: string google_hosted_domain = 3;
+   */
+  googleHostedDomain: string;
+
+  /**
+   * @generated from field: string google_user_id = 4;
+   */
+  googleUserId: string;
+
+  /**
+   * @generated from field: string microsoft_tenant_id = 5;
+   */
+  microsoftTenantId: string;
+
+  /**
+   * @generated from field: string microsoft_user_id = 6;
+   */
+  microsoftUserId: string;
+
+  /**
+   * @generated from field: bool revoked = 7;
+   */
+  revoked: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.IntermediateSession.
+ * Use `create(IntermediateSessionSchema)` to create a new message.
+ */
+export const IntermediateSessionSchema: GenMessage<IntermediateSession> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 10);
+
+/**
+ * @generated from message openauth.backend.v1.User
+ */
+export type User = Message<"openauth.backend.v1.User"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string organization_id = 2;
+   */
+  organizationId: string;
+
+  /**
+   * @generated from field: string email = 3;
+   */
+  email: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 4;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 5;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool owner = 6;
+   */
+  owner?: boolean;
+
+  /**
+   * @generated from field: optional string google_user_id = 7;
+   */
+  googleUserId?: string;
+
+  /**
+   * @generated from field: optional string microsoft_user_id = 8;
+   */
+  microsoftUserId?: string;
+
+  /**
+   * @generated from field: bool has_authenticator_app = 9;
+   */
+  hasAuthenticatorApp: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.User.
+ * Use `create(UserSchema)` to create a new message.
+ */
+export const UserSchema: GenMessage<User> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 11);
+
+/**
+ * @generated from message openauth.backend.v1.Session
+ */
+export type Session = Message<"openauth.backend.v1.Session"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string user_id = 2;
+   */
+  userId: string;
+
+  /**
+   * @generated from field: bool revoked = 3;
+   */
+  revoked: boolean;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 4;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp last_active_time = 5;
+   */
+  lastActiveTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp expire_time = 6;
+   */
+  expireTime?: Timestamp;
+};
+
+/**
+ * Describes the message openauth.backend.v1.Session.
+ * Use `create(SessionSchema)` to create a new message.
+ */
+export const SessionSchema: GenMessage<Session> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 12);
+
+/**
+ * @generated from message openauth.backend.v1.Passkey
+ */
+export type Passkey = Message<"openauth.backend.v1.Passkey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string user_id = 2;
+   */
+  userId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool disabled = 8;
+   */
+  disabled?: boolean;
+
+  /**
+   * @generated from field: bytes credential_id = 5;
+   */
+  credentialId: Uint8Array;
+
+  /**
+   * @generated from field: string public_key_pkix = 6;
+   */
+  publicKeyPkix: string;
+
+  /**
+   * @generated from field: string aaguid = 7;
+   */
+  aaguid: string;
+
+  /**
+   * @generated from field: string rp_id = 9;
+   */
+  rpId: string;
+};
+
+/**
+ * Describes the message openauth.backend.v1.Passkey.
+ * Use `create(PasskeySchema)` to create a new message.
+ */
+export const PasskeySchema: GenMessage<Passkey> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 13);
+
+/**
+ * @generated from message openauth.backend.v1.SAMLConnection
+ */
+export type SAMLConnection = Message<"openauth.backend.v1.SAMLConnection"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string organization_id = 2;
+   */
+  organizationId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool primary = 5;
+   */
+  primary?: boolean;
+
+  /**
+   * @generated from field: string sp_acs_url = 6;
+   */
+  spAcsUrl: string;
+
+  /**
+   * @generated from field: string sp_entity_id = 7;
+   */
+  spEntityId: string;
+
+  /**
+   * @generated from field: string idp_redirect_url = 8;
+   */
+  idpRedirectUrl: string;
+
+  /**
+   * @generated from field: string idp_x509_certificate = 9;
+   */
+  idpX509Certificate: string;
+
+  /**
+   * @generated from field: string idp_entity_id = 10;
+   */
+  idpEntityId: string;
+};
+
+/**
+ * Describes the message openauth.backend.v1.SAMLConnection.
+ * Use `create(SAMLConnectionSchema)` to create a new message.
+ */
+export const SAMLConnectionSchema: GenMessage<SAMLConnection> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 14);
+
+/**
+ * @generated from message openauth.backend.v1.SCIMAPIKey
+ */
+export type SCIMAPIKey = Message<"openauth.backend.v1.SCIMAPIKey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string organization_id = 2;
+   */
+  organizationId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: string display_name = 5;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: string secret_token = 6;
+   */
+  secretToken: string;
+
+  /**
+   * @generated from field: bool revoked = 7;
+   */
+  revoked: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.SCIMAPIKey.
+ * Use `create(SCIMAPIKeySchema)` to create a new message.
+ */
+export const SCIMAPIKeySchema: GenMessage<SCIMAPIKey> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 15);
+
+/**
+ * @generated from message openauth.backend.v1.UserImpersonationToken
+ */
+export type UserImpersonationToken = Message<"openauth.backend.v1.UserImpersonationToken"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string impersonator_id = 2;
+   */
+  impersonatorId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp expire_time = 4;
+   */
+  expireTime?: Timestamp;
+
+  /**
+   * @generated from field: string secret_token = 5;
+   */
+  secretToken: string;
+
+  /**
+   * @generated from field: string impersonated_id = 6;
+   */
+  impersonatedId: string;
+};
+
+/**
+ * Describes the message openauth.backend.v1.UserImpersonationToken.
+ * Use `create(UserImpersonationTokenSchema)` to create a new message.
+ */
+export const UserImpersonationTokenSchema: GenMessage<UserImpersonationToken> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 16);
+
+/**
+ * @generated from message openauth.backend.v1.UserInvite
+ */
+export type UserInvite = Message<"openauth.backend.v1.UserInvite"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string organization_id = 2;
+   */
+  organizationId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: string email = 5;
+   */
+  email: string;
+
+  /**
+   * @generated from field: bool owner = 6;
+   */
+  owner: boolean;
+};
+
+/**
+ * Describes the message openauth.backend.v1.UserInvite.
+ * Use `create(UserInviteSchema)` to create a new message.
+ */
+export const UserInviteSchema: GenMessage<UserInvite> = /*@__PURE__*/
+  messageDesc(file_openauth_backend_v1_models, 17);
+
