+++ internal/backend/proto/openauth/backend/v1/models.proto
@@ -1,0 +1,190 @@
+syntax = "proto3";
+
+package openauth.backend.v1;
+
+import "google/protobuf/timestamp.proto";
+import "google/protobuf/struct.proto";
+
+message Project {
+  string id = 1;
+  string display_name = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  optional bool log_in_with_google = 6;
+  optional bool log_in_with_microsoft = 7;
+  optional bool log_in_with_email = 15;
+  optional bool log_in_with_password = 5;
+  optional bool log_in_with_saml = 16;
+  optional bool log_in_with_authenticator_app = 13;
+  optional bool log_in_with_passkey = 14;
+  string google_oauth_client_id = 8;
+  string google_oauth_client_secret = 9;
+  string microsoft_oauth_client_id = 10;
+  string microsoft_oauth_client_secret = 11;
+  optional string auth_domain = 12;
+  repeated string trusted_domains = 17;
+  string redirect_uri = 18;
+  string after_login_redirect_uri = 19;
+  string after_signup_redirect_uri = 20;
+}
+
+message VaultDomainSettings {
+  string pending_domain = 1;
+  VaultDomainSettingsDNSRecord tesseral_cname_record = 2;
+  bool tesseral_cname_record_configured = 3;
+  repeated VaultDomainSettingsDNSRecord dkim_records = 4;
+  bool dkim_configured = 5;
+  repeated VaultDomainSettingsDNSRecord spf_records = 6;
+  bool spf_configured = 7;
+}
+
+message VaultDomainSettingsDNSRecord {
+  string type = 1;
+  string name = 2;
+  string value = 3;
+}
+
+message ProjectUISettings {
+  string id = 1;
+  string project_id = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  string logo_url = 5;
+  string favicon_url = 6;
+  string primary_color = 7;
+  bool detect_dark_mode_enabled = 8;
+  string dark_mode_logo_url = 9;
+  string dark_mode_primary_color = 10;
+  string log_in_layout = 11;
+}
+
+message Organization {
+  string id = 1;
+  string display_name = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  optional bool log_in_with_google = 7;
+  optional bool log_in_with_microsoft = 8;
+  optional bool log_in_with_email = 14;
+  optional bool log_in_with_password = 6;
+  optional bool log_in_with_saml = 9;
+  optional bool log_in_with_authenticator_app = 11;
+  optional bool log_in_with_passkey = 12;
+  optional bool require_mfa = 13;
+  optional bool scim_enabled = 10;
+}
+
+message OrganizationGoogleHostedDomains {
+  string organization_id = 1;
+  repeated string google_hosted_domains = 2;
+}
+
+message OrganizationMicrosoftTenantIDs {
+  string organization_id = 1;
+  repeated string microsoft_tenant_ids = 2;
+}
+
+message SessionSigningKey {
+  string id = 1;
+  google.protobuf.Struct public_key_jwk = 2;
+}
+
+message ProjectAPIKey {
+  string id = 1;
+  string display_name = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  string secret_token = 5;
+  bool revoked = 6;
+}
+
+message PublishableKey {
+  string id = 1;
+  string display_name = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+}
+
+message IntermediateSession {
+  string id = 1;
+  string email = 2;
+  string google_hosted_domain = 3;
+  string google_user_id = 4;
+  string microsoft_tenant_id = 5;
+  string microsoft_user_id = 6;
+  bool revoked = 7;
+}
+
+message User {
+  string id = 1;
+  string organization_id = 2;
+  string email = 3;
+  google.protobuf.Timestamp create_time = 4;
+  google.protobuf.Timestamp update_time = 5;
+  optional bool owner = 6;
+  optional string google_user_id = 7;
+  optional string microsoft_user_id = 8;
+  bool has_authenticator_app = 9;
+}
+
+message Session {
+  string id = 1;
+  string user_id = 2;
+  bool revoked = 3;
+  google.protobuf.Timestamp create_time = 4;
+  google.protobuf.Timestamp last_active_time = 5;
+  google.protobuf.Timestamp expire_time = 6;
+}
+
+message Passkey {
+  string id = 1;
+  string user_id = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  optional bool disabled = 8;
+  bytes credential_id = 5;
+  string public_key_pkix = 6;
+  string aaguid = 7;
+  string rp_id = 9;
+}
+
+message SAMLConnection {
+  string id = 1;
+  string organization_id = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  optional bool primary = 5;
+  string sp_acs_url = 6;
+  string sp_entity_id = 7;
+  string idp_redirect_url = 8;
+  string idp_x509_certificate = 9;
+  string idp_entity_id = 10;
+}
+
+message SCIMAPIKey {
+  string id = 1;
+  string organization_id = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  string display_name = 5;
+  string secret_token = 6;
+  bool revoked = 7;
+}
+
+message UserImpersonationToken {
+  string id = 1;
+  string impersonator_id = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp expire_time = 4;
+  string secret_token = 5;
+  string impersonated_id = 6;
+}
+
+message UserInvite {
+  string id = 1;
+  string organization_id = 2;
+  google.protobuf.Timestamp create_time = 3;
+  google.protobuf.Timestamp update_time = 4;
+  string email = 5;
+  bool owner = 6;
+}
