+++ ui/src/gen/openauth/frontend/v1/models_pb.ts
@@ -1,0 +1,513 @@
+// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
+// @generated from file openauth/frontend/v1/models.proto (package openauth.frontend.v1, syntax proto3)
+/* eslint-disable */
+
+import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
+import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
+import type { Timestamp } from "@bufbuild/protobuf/wkt";
+import { file_google_protobuf_struct, file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
+import type { JsonObject, Message } from "@bufbuild/protobuf";
+
+/**
+ * Describes the file openauth/frontend/v1/models.proto.
+ */
+export const file_openauth_frontend_v1_models: GenFile = /*@__PURE__*/
+  fileDesc("CiFvcGVuYXV0aC9mcm9udGVuZC92MS9tb2RlbHMucHJvdG8SFG9wZW5hdXRoLmZyb250ZW5kLnYxIvQCCgdQcm9qZWN0EgoKAmlkGAEgASgJEhQKDGRpc3BsYXlfbmFtZRgCIAEoCRIvCgtjcmVhdGVfdGltZRgDIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYBCABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEhoKEmxvZ19pbl93aXRoX2dvb2dsZRgGIAEoCBIdChVsb2dfaW5fd2l0aF9taWNyb3NvZnQYByABKAgSGQoRbG9nX2luX3dpdGhfZW1haWwYCyABKAgSHAoUbG9nX2luX3dpdGhfcGFzc3dvcmQYBSABKAgSGAoQbG9nX2luX3dpdGhfc2FtbBgMIAEoCBIlCh1sb2dfaW5fd2l0aF9hdXRoZW50aWNhdG9yX2FwcBgJIAEoCBIbChNsb2dfaW5fd2l0aF9wYXNza2V5GAogASgIEhMKC2F1dGhfZG9tYWluGAggASgJIrMFCgxPcmdhbml6YXRpb24SCgoCaWQYASABKAkSFAoMZGlzcGxheV9uYW1lGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgt1cGRhdGVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASHwoSbG9nX2luX3dpdGhfZ29vZ2xlGAcgASgISACIAQESIgoVbG9nX2luX3dpdGhfbWljcm9zb2Z0GAggASgISAGIAQESHgoRbG9nX2luX3dpdGhfZW1haWwYECABKAhIAogBARIhChRsb2dfaW5fd2l0aF9wYXNzd29yZBgGIAEoCEgDiAEBEh0KEGxvZ19pbl93aXRoX3NhbWwYDCABKAhIBIgBARIqCh1sb2dfaW5fd2l0aF9hdXRoZW50aWNhdG9yX2FwcBgNIAEoCEgFiAEBEiAKE2xvZ19pbl93aXRoX3Bhc3NrZXkYDiABKAhIBogBARIYCgtyZXF1aXJlX21mYRgPIAEoCEgHiAEBEh0KFWdvb2dsZV9ob3N0ZWRfZG9tYWlucxgJIAMoCRIcChRtaWNyb3NvZnRfdGVuYW50X2lkcxgKIAMoCRIUCgxzYW1sX2VuYWJsZWQYCyABKAhCFQoTX2xvZ19pbl93aXRoX2dvb2dsZUIYChZfbG9nX2luX3dpdGhfbWljcm9zb2Z0QhQKEl9sb2dfaW5fd2l0aF9lbWFpbEIXChVfbG9nX2luX3dpdGhfcGFzc3dvcmRCEwoRX2xvZ19pbl93aXRoX3NhbWxCIAoeX2xvZ19pbl93aXRoX2F1dGhlbnRpY2F0b3JfYXBwQhYKFF9sb2dfaW5fd2l0aF9wYXNza2V5Qg4KDF9yZXF1aXJlX21mYSJACh9Pcmdhbml6YXRpb25Hb29nbGVIb3N0ZWREb21haW5zEh0KFWdvb2dsZV9ob3N0ZWRfZG9tYWlucxgCIAMoCSI+Ch5Pcmdhbml6YXRpb25NaWNyb3NvZnRUZW5hbnRJRHMSHAoUbWljcm9zb2Z0X3RlbmFudF9pZHMYAiADKAkiUAoRU2Vzc2lvblNpZ25pbmdLZXkSCgoCaWQYASABKAkSLwoOcHVibGljX2tleV9qd2sYAiABKAsyFy5nb29nbGUucHJvdG9idWYuU3RydWN0IvMBCgRVc2VyEgoKAmlkGAEgASgJEg0KBWVtYWlsGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgt1cGRhdGVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASEgoFb3duZXIYBSABKAhIAIgBARIWCg5nb29nbGVfdXNlcl9pZBgGIAEoCRIZChFtaWNyb3NvZnRfdXNlcl9pZBgHIAEoCRIdChVoYXNfYXV0aGVudGljYXRvcl9hcHAYCCABKAhCCAoGX293bmVyIpkBCgdTZXNzaW9uEgoKAmlkGAEgASgJEg8KB3VzZXJfaWQYAiABKAkSLwoLY3JlYXRlX3RpbWUYAyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEi8KC2V4cGlyZV90aW1lGAQgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIPCgdyZXZva2VkGAUgASgIIpkCCg5TQU1MQ29ubmVjdGlvbhIKCgJpZBgBIAEoCRIvCgtjcmVhdGVfdGltZRgCIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYAyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEhQKB3ByaW1hcnkYBCABKAhIAIgBARISCgpzcF9hY3NfdXJsGAUgASgJEhQKDHNwX2VudGl0eV9pZBgGIAEoCRIYChBpZHBfcmVkaXJlY3RfdXJsGAcgASgJEhwKFGlkcF94NTA5X2NlcnRpZmljYXRlGAggASgJEhUKDWlkcF9lbnRpdHlfaWQYCSABKAlCCgoIX3ByaW1hcnkitwEKClNDSU1BUElLZXkSCgoCaWQYASABKAkSFAoMZGlzcGxheV9uYW1lGAIgASgJEi8KC2NyZWF0ZV90aW1lGAMgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIvCgt1cGRhdGVfdGltZRgEIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASFAoMc2VjcmV0X3Rva2VuGAUgASgJEg8KB3Jldm9rZWQYBiABKAgi6QEKB1Bhc3NrZXkSCgoCaWQYASABKAkSDwoHdXNlcl9pZBgCIAEoCRIvCgtjcmVhdGVfdGltZRgDIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYBCABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEhAKCGRpc2FibGVkGAggASgIEhUKDWNyZWRlbnRpYWxfaWQYBSABKAwSFwoPcHVibGljX2tleV9wa2l4GAYgASgJEg4KBmFhZ3VpZBgHIAEoCRINCgVycF9pZBgJIAEoCSKYAQoKVXNlckludml0ZRIKCgJpZBgBIAEoCRIvCgtjcmVhdGVfdGltZRgCIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXASLwoLdXBkYXRlX3RpbWUYAyABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEg0KBWVtYWlsGAQgASgJEg0KBW93bmVyGAUgASgIQu0BChhjb20ub3BlbmF1dGguZnJvbnRlbmQudjFCC01vZGVsc1Byb3RvUAFaUmdpdGh1Yi5jb20vb3BlbmF1dGgvb3BlbmF1dGgvaW50ZXJuYWwvZnJvbnRlbmQvZ2VuL29wZW5hdXRoL2Zyb250ZW5kL3YxO2Zyb250ZW5kdjGiAgNPRliqAhRPcGVuYXV0aC5Gcm9udGVuZC5WMcoCFE9wZW5hdXRoXEZyb250ZW5kXFYx4gIgT3BlbmF1dGhcRnJvbnRlbmRcVjFcR1BCTWV0YWRhdGHqAhZPcGVuYXV0aDo6RnJvbnRlbmQ6OlYxYgZwcm90bzM", [file_google_protobuf_timestamp, file_google_protobuf_struct]);
+
+/**
+ * @generated from message openauth.frontend.v1.Project
+ */
+export type Project = Message<"openauth.frontend.v1.Project"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: bool log_in_with_google = 6;
+   */
+  logInWithGoogle: boolean;
+
+  /**
+   * @generated from field: bool log_in_with_microsoft = 7;
+   */
+  logInWithMicrosoft: boolean;
+
+  /**
+   * @generated from field: bool log_in_with_email = 11;
+   */
+  logInWithEmail: boolean;
+
+  /**
+   * @generated from field: bool log_in_with_password = 5;
+   */
+  logInWithPassword: boolean;
+
+  /**
+   * @generated from field: bool log_in_with_saml = 12;
+   */
+  logInWithSaml: boolean;
+
+  /**
+   * @generated from field: bool log_in_with_authenticator_app = 9;
+   */
+  logInWithAuthenticatorApp: boolean;
+
+  /**
+   * @generated from field: bool log_in_with_passkey = 10;
+   */
+  logInWithPasskey: boolean;
+
+  /**
+   * @generated from field: string auth_domain = 8;
+   */
+  authDomain: string;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.Project.
+ * Use `create(ProjectSchema)` to create a new message.
+ */
+export const ProjectSchema: GenMessage<Project> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 0);
+
+/**
+ * @generated from message openauth.frontend.v1.Organization
+ */
+export type Organization = Message<"openauth.frontend.v1.Organization"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool log_in_with_google = 7;
+   */
+  logInWithGoogle?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_microsoft = 8;
+   */
+  logInWithMicrosoft?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_email = 16;
+   */
+  logInWithEmail?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_password = 6;
+   */
+  logInWithPassword?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_saml = 12;
+   */
+  logInWithSaml?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_authenticator_app = 13;
+   */
+  logInWithAuthenticatorApp?: boolean;
+
+  /**
+   * @generated from field: optional bool log_in_with_passkey = 14;
+   */
+  logInWithPasskey?: boolean;
+
+  /**
+   * @generated from field: optional bool require_mfa = 15;
+   */
+  requireMfa?: boolean;
+
+  /**
+   * @generated from field: repeated string google_hosted_domains = 9;
+   */
+  googleHostedDomains: string[];
+
+  /**
+   * @generated from field: repeated string microsoft_tenant_ids = 10;
+   */
+  microsoftTenantIds: string[];
+
+  /**
+   * @generated from field: bool saml_enabled = 11;
+   */
+  samlEnabled: boolean;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.Organization.
+ * Use `create(OrganizationSchema)` to create a new message.
+ */
+export const OrganizationSchema: GenMessage<Organization> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 1);
+
+/**
+ * @generated from message openauth.frontend.v1.OrganizationGoogleHostedDomains
+ */
+export type OrganizationGoogleHostedDomains = Message<"openauth.frontend.v1.OrganizationGoogleHostedDomains"> & {
+  /**
+   * @generated from field: repeated string google_hosted_domains = 2;
+   */
+  googleHostedDomains: string[];
+};
+
+/**
+ * Describes the message openauth.frontend.v1.OrganizationGoogleHostedDomains.
+ * Use `create(OrganizationGoogleHostedDomainsSchema)` to create a new message.
+ */
+export const OrganizationGoogleHostedDomainsSchema: GenMessage<OrganizationGoogleHostedDomains> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 2);
+
+/**
+ * @generated from message openauth.frontend.v1.OrganizationMicrosoftTenantIDs
+ */
+export type OrganizationMicrosoftTenantIDs = Message<"openauth.frontend.v1.OrganizationMicrosoftTenantIDs"> & {
+  /**
+   * @generated from field: repeated string microsoft_tenant_ids = 2;
+   */
+  microsoftTenantIds: string[];
+};
+
+/**
+ * Describes the message openauth.frontend.v1.OrganizationMicrosoftTenantIDs.
+ * Use `create(OrganizationMicrosoftTenantIDsSchema)` to create a new message.
+ */
+export const OrganizationMicrosoftTenantIDsSchema: GenMessage<OrganizationMicrosoftTenantIDs> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 3);
+
+/**
+ * @generated from message openauth.frontend.v1.SessionSigningKey
+ */
+export type SessionSigningKey = Message<"openauth.frontend.v1.SessionSigningKey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: google.protobuf.Struct public_key_jwk = 2;
+   */
+  publicKeyJwk?: JsonObject;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.SessionSigningKey.
+ * Use `create(SessionSigningKeySchema)` to create a new message.
+ */
+export const SessionSigningKeySchema: GenMessage<SessionSigningKey> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 4);
+
+/**
+ * @generated from message openauth.frontend.v1.User
+ */
+export type User = Message<"openauth.frontend.v1.User"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string email = 2;
+   */
+  email: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool owner = 5;
+   */
+  owner?: boolean;
+
+  /**
+   * @generated from field: string google_user_id = 6;
+   */
+  googleUserId: string;
+
+  /**
+   * @generated from field: string microsoft_user_id = 7;
+   */
+  microsoftUserId: string;
+
+  /**
+   * @generated from field: bool has_authenticator_app = 8;
+   */
+  hasAuthenticatorApp: boolean;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.User.
+ * Use `create(UserSchema)` to create a new message.
+ */
+export const UserSchema: GenMessage<User> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 5);
+
+/**
+ * @generated from message openauth.frontend.v1.Session
+ */
+export type Session = Message<"openauth.frontend.v1.Session"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string user_id = 2;
+   */
+  userId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp expire_time = 4;
+   */
+  expireTime?: Timestamp;
+
+  /**
+   * @generated from field: bool revoked = 5;
+   */
+  revoked: boolean;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.Session.
+ * Use `create(SessionSchema)` to create a new message.
+ */
+export const SessionSchema: GenMessage<Session> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 6);
+
+/**
+ * @generated from message openauth.frontend.v1.SAMLConnection
+ */
+export type SAMLConnection = Message<"openauth.frontend.v1.SAMLConnection"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 2;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 3;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: optional bool primary = 4;
+   */
+  primary?: boolean;
+
+  /**
+   * @generated from field: string sp_acs_url = 5;
+   */
+  spAcsUrl: string;
+
+  /**
+   * @generated from field: string sp_entity_id = 6;
+   */
+  spEntityId: string;
+
+  /**
+   * @generated from field: string idp_redirect_url = 7;
+   */
+  idpRedirectUrl: string;
+
+  /**
+   * @generated from field: string idp_x509_certificate = 8;
+   */
+  idpX509Certificate: string;
+
+  /**
+   * @generated from field: string idp_entity_id = 9;
+   */
+  idpEntityId: string;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.SAMLConnection.
+ * Use `create(SAMLConnectionSchema)` to create a new message.
+ */
+export const SAMLConnectionSchema: GenMessage<SAMLConnection> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 7);
+
+/**
+ * @generated from message openauth.frontend.v1.SCIMAPIKey
+ */
+export type SCIMAPIKey = Message<"openauth.frontend.v1.SCIMAPIKey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string display_name = 2;
+   */
+  displayName: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: string secret_token = 5;
+   */
+  secretToken: string;
+
+  /**
+   * @generated from field: bool revoked = 6;
+   */
+  revoked: boolean;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.SCIMAPIKey.
+ * Use `create(SCIMAPIKeySchema)` to create a new message.
+ */
+export const SCIMAPIKeySchema: GenMessage<SCIMAPIKey> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 8);
+
+/**
+ * @generated from message openauth.frontend.v1.Passkey
+ */
+export type Passkey = Message<"openauth.frontend.v1.Passkey"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: string user_id = 2;
+   */
+  userId: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 3;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 4;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: bool disabled = 8;
+   */
+  disabled: boolean;
+
+  /**
+   * @generated from field: bytes credential_id = 5;
+   */
+  credentialId: Uint8Array;
+
+  /**
+   * @generated from field: string public_key_pkix = 6;
+   */
+  publicKeyPkix: string;
+
+  /**
+   * @generated from field: string aaguid = 7;
+   */
+  aaguid: string;
+
+  /**
+   * @generated from field: string rp_id = 9;
+   */
+  rpId: string;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.Passkey.
+ * Use `create(PasskeySchema)` to create a new message.
+ */
+export const PasskeySchema: GenMessage<Passkey> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 9);
+
+/**
+ * @generated from message openauth.frontend.v1.UserInvite
+ */
+export type UserInvite = Message<"openauth.frontend.v1.UserInvite"> & {
+  /**
+   * @generated from field: string id = 1;
+   */
+  id: string;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp create_time = 2;
+   */
+  createTime?: Timestamp;
+
+  /**
+   * @generated from field: google.protobuf.Timestamp update_time = 3;
+   */
+  updateTime?: Timestamp;
+
+  /**
+   * @generated from field: string email = 4;
+   */
+  email: string;
+
+  /**
+   * @generated from field: bool owner = 5;
+   */
+  owner: boolean;
+};
+
+/**
+ * Describes the message openauth.frontend.v1.UserInvite.
+ * Use `create(UserInviteSchema)` to create a new message.
+ */
+export const UserInviteSchema: GenMessage<UserInvite> = /*@__PURE__*/
+  messageDesc(file_openauth_frontend_v1_models, 10);
+
